ID3v2: Move `Extended{Text,Url}Frame` parse functions to structs
