+++ src/id3/v2/items/extended_text_frame.rs
@@ -1,6 +1,12 @@
-use crate::util::text::{encode_text, TextEncoding};
+use crate::error::{ID3v2Error, ID3v2ErrorKind, LoftyError, Result};
+use crate::id3::v2::frame::content::verify_encoding;
+use crate::id3::v2::ID3v2Version;
+use crate::util::text::{decode_text, encode_text, read_to_terminator, utf16_decode, TextEncoding};
 
 use std::hash::{Hash, Hasher};
+use std::io::{Cursor, Read};
+
+use byteorder::ReadBytesExt;
 
 /// An extended `ID3v2` text frame
 ///
@@ -31,6 +37,68 @@ impl Hash for ExtendedTextFrame {
 }
 
 impl ExtendedTextFrame {
+	/// Read an [`ExtendedTextFrame`] from a slice
+	///
+	/// NOTE: This expects the frame header to have already been skipped
+	///
+	/// # Errors
+	///
+	/// * Unable to decode the text
+	///
+	/// ID3v2.2:
+	///
+	/// * The encoding is not [`TextEncoding::Latin1`] or [`TextEncoding::UTF16`]
+	pub fn parse(content: &[u8], version: ID3v2Version) -> Result<Option<Self>> {
+		if content.len() < 2 {
+			return Ok(None);
+		}
+
+		let mut content = &mut &content[..];
+		let encoding = verify_encoding(content.read_u8()?, version)?;
+
+		let mut endianness: fn([u8; 2]) -> u16 = u16::from_le_bytes;
+		if encoding == TextEncoding::UTF16 {
+			let mut cursor = Cursor::new(content);
+			let mut bom = [0; 2];
+			cursor.read_exact(&mut bom)?;
+
+			match [bom[0], bom[1]] {
+				[0xFF, 0xFE] => endianness = u16::from_le_bytes,
+				[0xFE, 0xFF] => endianness = u16::from_be_bytes,
+				// We'll catch an invalid BOM below
+				_ => {},
+			};
+
+			content = cursor.into_inner();
+		}
+
+		let description = decode_text(content, encoding, true)?.unwrap_or_default();
+
+		let frame_content;
+		// It's possible for the description to be the only string with a BOM
+		if encoding == TextEncoding::UTF16 {
+			if content.len() >= 2 && (content[..2] == [0xFF, 0xFE] || content[..2] == [0xFE, 0xFF])
+			{
+				frame_content = decode_text(content, encoding, false)?.unwrap_or_default();
+			} else {
+				frame_content = match read_to_terminator(content, TextEncoding::UTF16) {
+					Some(raw_text) => utf16_decode(&raw_text, endianness).map_err(|_| {
+						Into::<LoftyError>::into(ID3v2Error::new(ID3v2ErrorKind::BadSyncText))
+					})?,
+					None => String::new(),
+				}
+			}
+		} else {
+			frame_content = decode_text(content, encoding, false)?.unwrap_or_default();
+		}
+
+		Ok(Some(ExtendedTextFrame {
+			encoding,
+			description,
+			content: frame_content,
+		}))
+	}
+
 	/// Convert an [`ExtendedTextFrame`] to a byte vec
 	pub fn as_bytes(&self) -> Vec<u8> {
 		let mut bytes = vec![self.encoding as u8];
